requirements.txt
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
Faker==25.3.0
gunicorn==22.0.0
uvicorn==0.30.0
